+++ Calculadora/src/main/java/br/com/Calculadora/Service/MedicamentoService.java
@@ -8,6 +8,7 @@ import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import br.com.Calculadora.Dto.GrupoMedicamentoDto;
@@ -33,22 +34,35 @@ public class MedicamentoService {
 		return (medicamento);
 	}
 
-	public ResponseEntity<MedicamentoDto> criar(@Valid MedicamentoForm medicamentoForm,
-			UriComponentsBuilder uriBuilder) {
-		Medicamento medicamento = medicamentoForm.converter();
-		medicamentoRepository.save(medicamento);
+	public ResponseEntity<MedicamentoDto> criar(@Valid MedicamentoForm medicamentoForm, UriComponentsBuilder uriBuilder,
+			BindingResult result) {
+		if (result.hasErrors()) {
+			throw new RuntimeException();
+		} else {
+			Medicamento medicamento = medicamentoForm.converter();
+			medicamentoRepository.save(medicamento);
 
-		URI uri = uriBuilder.path("/criar/{id}").buildAndExpand(medicamento.getId()).toUri();
-		return ResponseEntity.created(uri).body(new MedicamentoDto(medicamento));
+			URI uri = uriBuilder.path("/criar/{id}").buildAndExpand(medicamento.getId()).toUri();
+			return ResponseEntity.created(uri).body(new MedicamentoDto(medicamento));
+		}
 	}
 
 	public ResponseEntity<MedicamentoDto> remover(BigInteger id) {
-		medicamentoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		if (!medicamentoRepository.existsById(id)) {
+			throw new RuntimeException();
+		} else {
+			medicamentoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
 	}
-	
-	public ResponseEntity<MedicamentoDto> atualizar(BigInteger id, MedicamentoForm medicamentoForm){
-		Medicamento medicamento = medicamentoForm.atualizar(id, medicamentoRepository);
-		return(ResponseEntity.ok(new MedicamentoDto(medicamento)));
+
+	public ResponseEntity<MedicamentoDto> atualizar(BigInteger id, MedicamentoForm medicamentoForm,
+			BindingResult result) {
+		if (result.hasErrors() || !medicamentoRepository.existsById(id)) {
+			throw new RuntimeException();
+		} else {
+			Medicamento medicamento = medicamentoForm.atualizar(id, medicamentoRepository);
+			return (ResponseEntity.ok(new MedicamentoDto(medicamento)));
+		}
 	}
 }
